*****IMPORTANT****
This is an IMPORTANT commit.
The pop-up when clicking on the result cards can populate exact GalleryImgs as defined in the card/result/__WebCCLVietnameseResultsCardData.json
which corresponds to the image shows (Not by matching ID) but by double mapping the cardData and the GalleryImgs.

In WebCCLVietnameseTabCardGrid.js, once a card is selected, selectedCardData and setIsModalOpen are hooked via useState.
The TabData arrays are then passed to the GallerySlider to be mapped and displayed.

 This pop-up is designed to display a set of images ("GalleryImgs") which are defined in the __WebCCLVietnameseResultsCardData.json file.

These images don't correspond to the ones shown by matching their IDs. Instead, you've utilized a process of double mapping involving the cardData and GalleryImgs.

In your WebCCLVietnameseTabCardGrid.js file, you've used a hook useState for selectedCardData and setIsModalOpen when a card is selected. This state manages the display of the modal and the data it contains.

The arrays of TabData are then passed on to GallerySlider to be mapped and displayed in the pop-up.